[gateway]: fix problems along with time size from uint64 to uint16

+++ src/cluster/gateway/gateway.go
@@ -28,8 +28,8 @@ type Config struct {
 	ClusterGroup      string
 	ClusterMemberMode Mode
 
-	OPLogMaxSeqGapToPull  uint64
-	OPLogPullMaxCountOnce uint64
+	OPLogMaxSeqGapToPull  uint16
+	OPLogPullMaxCountOnce uint16
 	OPLogPullInterval     time.Duration
 	OPLogPullTimeout      time.Duration
 }
